[cleanup] CustomInterface refactored out of action
